+++ ghost/core/core/server/services/email-analytics/EmailAnalyticsProviderMailgun.js
@@ -1,4 +1,4 @@
-const MailgunClient = require('@tryghost/mailgun-client');
+const MailgunClient = require('../lib/MailgunClient');
 
 const DEFAULT_EVENT_FILTER = 'delivered OR opened OR failed OR unsubscribed OR complained';
 const PAGE_LIMIT = 300;
