+++ ghost/core/core/server/services/email-service/EmailServiceWrapper.js
@@ -17,7 +17,7 @@ class EmailServiceWrapper {
 
         const {EmailService, EmailController, EmailRenderer, SendingService, BatchSendingService, EmailSegmenter, MailgunEmailProvider} = require('@tryghost/email-service');
         const {Post, Newsletter, Email, EmailBatch, EmailRecipient, Member} = require('../../models');
-        const MailgunClient = require('@tryghost/mailgun-client');
+        const MailgunClient = require('../lib/MailgunClient');
         const configService = require('../../../shared/config');
         const settingsCache = require('../../../shared/settings-cache');
         const settingsHelpers = require('../settings-helpers');
@@ -53,7 +53,7 @@ class EmailServiceWrapper {
         });
         const i18nLanguage = labs.isSet('i18n') ? settingsCache.get('locale') || 'en' : 'en';
         const i18n = i18nLib(i18nLanguage, 'newsletter');
-        
+
         events.on('settings.labs.edited', () => {
             if (labs.isSet('i18n')) {
                 debug('labs i18n enabled, updating i18n to', settingsCache.get('locale'));
@@ -63,12 +63,12 @@ class EmailServiceWrapper {
                 i18n.changeLanguage('en');
             }
         });
-    
+
         events.on('settings.locale.edited', (model) => {
             if (labs.isSet('i18n')) {
                 debug('locale changed, updating i18n to', model.get('value'));
                 i18n.changeLanguage(model.get('value'));
-            } 
+            }
         });
 
         const mailgunEmailProvider = new MailgunEmailProvider({
