+++ ghost/core/package.json
@@ -103,7 +103,6 @@
     "@tryghost/link-replacer": "0.0.0",
     "@tryghost/logging": "2.4.21",
     "@tryghost/magic-link": "0.0.0",
-    "@tryghost/mailgun-client": "0.0.0",
     "@tryghost/member-attribution": "0.0.0",
     "@tryghost/member-events": "0.0.0",
     "@tryghost/members-api": "0.0.0",
@@ -160,6 +159,7 @@
     "express-queue": "0.0.13",
     "express-session": "1.18.1",
     "file-type": "16.5.4",
+    "form-data": "4.0.0",
     "fs-extra": "11.3.0",
     "ghost-storage-base": "1.0.0",
     "glob": "8.1.0",
@@ -184,6 +184,7 @@
     "leaky-bucket": "2.2.0",
     "lodash": "4.17.21",
     "luxon": "3.6.1",
+    "mailgun.js": "10.4.0",
     "metascraper": "5.45.15",
     "metascraper-author": "5.45.10",
     "metascraper-description": "5.45.10",
