+++ ghost/core/test/e2e-browser/fixtures/ghost-test.js
@@ -4,11 +4,11 @@ const {promisify} = require('util');
 const {spawn, exec} = require('child_process');
 const {setupGhost, setupMailgun, enableLabs, setupStripe, getStripeAccountId, generateStripeIntegrationToken} = require('../utils/e2e-browser-utils');
 const {allowStripe, mockMail, mockGeojs, assert} = require('../../utils/e2e-framework-mock-manager');
-const MailgunClient = require('@tryghost/mailgun-client');
 const sinon = require('sinon');
 const ObjectID = require('bson-objectid').default;
 const Stripe = require('stripe').Stripe;
 const configUtils = require('../../utils/configUtils');
+const MailgunClient = require('../../../core/server/services/lib/MailgunClient');
 
 const startWebhookServer = (port) => {
     const isCI = process.env.CI;
