+++ ghost/core/test/integration/services/email-service/email-event-storage.test.js
@@ -1,7 +1,7 @@
 const sinon = require('sinon');
 const {agentProvider, fixtureManager} = require('../../../utils/e2e-framework');
 const assert = require('assert/strict');
-const MailgunClient = require('@tryghost/mailgun-client');
+const MailgunClient = require('../../../../core/server/services/lib/MailgunClient');
 const DomainEvents = require('@tryghost/domain-events');
 const emailAnalytics = require('../../../../core/server/services/email-analytics');
