+++ ghost/core/test/integration/services/mailgun-email-suppression-list.test.js
@@ -1,8 +1,9 @@
 const sinon = require('sinon');
 const {agentProvider, fixtureManager} = require('../../utils/e2e-framework');
 const assert = require('assert/strict');
-const MailgunClient = require('@tryghost/mailgun-client');
 const DomainEvents = require('@tryghost/domain-events');
+
+const MailgunClient = require('../../../core/server/services/lib/MailgunClient');
 const emailAnalytics = require('../../../core/server/services/email-analytics');
 
 describe('MailgunEmailSuppressionList', function () {
