--- ghost/mailgun-client/README.md
@@ -1,11 +1,0 @@
-# Mailgun Client
-
-Mailgun client for Ghost
-
-## Usage
-
-## Test
-
-- `yarn lint` run just eslint
-- `yarn test` run lint and tests
-
