--- ghost/mailgun-client/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@tryghost/mailgun-client",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/mailgun-client",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "nock": "13.5.6",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/debug": "0.1.32",
-    "@tryghost/logging": "2.4.21",
-    "@tryghost/metrics": "1.0.37",
-    "form-data": "4.0.0",
-    "lodash": "4.17.21",
-    "mailgun.js": "10.4.0"
-  }
-}
